replaced music.cpp with music_mix.cpp
