+++ src/music.cpp
@@ -23,155 +23,120 @@
  * \brief In-game music routines
  *
  * Handles playing and pausing music in the game.
- * Interfaces to DUMB
+ * Interfaces to SDL2_Mixer
  */
 
-#include <cstdio>
-#include <cstring>
-#include <string>
-
 #include "disk.h"
+#include "imgcache.h"
 #include "kq.h"
 #include "music.h"
 #include "platform.h"
+#include <SDL_mixer.h>
+#include <algorithm>
+#include <string>
 
-/* DUMB version of music */
-#include <dumb.h>
+/* SDL version of music */
+struct Mix_MusicLoader
+{
+    Mix_Music* operator()(const std::string&);
+};
+struct Mix_MusicDeleter
+{
+    void operator()(Mix_Music* m)
+    {
+        Mix_FreeMusic(m);
+    }
+};
+struct Mix_ChunkLoader
+{
+    Mix_Chunk* operator()(const std::string&);
+};
+struct Mix_ChunkDeleter
+{
+    void operator()(Mix_Chunk*);
+};
 
-/* private variables */
-#define MAX_MUSIC_PLAYERS 3
-static DUH* mod_song[MAX_MUSIC_PLAYERS];
-static void* mod_player[MAX_MUSIC_PLAYERS];
-static int current_music_player;
+static Cache<Mix_Chunk, Mix_ChunkLoader, Mix_ChunkDeleter> sample_cache;
 
-/*! \brief Initiate music player (DUMB)
+static Cache<Mix_Music, Mix_MusicLoader, Mix_MusicDeleter> music_cache;
+static Mix_Music* music = nullptr;
+
+/*! \brief Initiate music player (SDL2_Mixer)
  *
  * Initializes the music players. Must be called before any other
  * music function. Needs to be shutdown when finished.
  */
 void KMusic::init_music(void)
 {
-    atexit(&dumb_exit);
-    dumb_register_stdfiles();
-    dumb_resampling_quality = 2;
-
-    /* initialize all music players */
-    current_music_player = MAX_MUSIC_PLAYERS;
-    while (current_music_player--)
+    int rc = Mix_OpenAudio(MIX_DEFAULT_FREQUENCY, MIX_DEFAULT_FORMAT, 2, 1024);
+    if (rc == -1)
     {
-        mod_song[current_music_player] = NULL;
-        mod_player[current_music_player] = NULL;
+        Game.program_death("Could not open audio: %s", Mix_GetError());
     }
-    current_music_player = 0;
 }
 
-/*! \brief Clean up and shut down music (DUMB)
+/*! \brief Clean up and shut down music  (SDL2_Mixer)
  *
  * Performs any cleanup needed. Must be called before the program exits.
  */
 void KMusic::shutdown_music(void)
 {
-    if (is_sound != 0)
-    {
-        do
-        {
-            stop_music();
-        } while (current_music_player--);
-    }
+    music_cache.clear();
+    Mix_CloseAudio();
 }
 
-/*! \brief Set the music volume (DUMB)
+/*! \brief Set the music volume (SDL2_Mixer)
  *
  * Sets the volume of the currently playing music.
  *
- * \param   volume 0 (silent) to 100 (loudest)
+ * \param   volume 0 (silent) to 250 (loudest)
  */
-void KMusic::set_music_volume(float volume)
+void KMusic::set_music_volume(int volume)
 {
-    if (is_sound != 0 && mod_player[current_music_player])
-    {
-        /* TODO
-          al_duh_set_volume(mod_player[current_music_player], volume);
-        */
-    }
+    mvol = float(std::clamp(volume, 0, 250)) / 250.0f;
+    Mix_VolumeMusic(int(dvol * mvol * float(MIX_MAX_VOLUME)));
 }
 
-/*! \brief Poll the music (DUMB)
+/*! \brief Poll the music  (SDL2_Mixer)
  *
  * Does whatever is needed to ensure the music keeps playing.
  * It's safe to call this too much, but shouldn't be called inside a timer.
  */
 void KMusic::poll_music(void)
 {
-    if (is_sound != 0)
-    {
-        /* TODO
-          al_poll_duh(mod_player[current_music_player]);
-        */
-    }
+    // No-op for SDL_mixer
 }
 
-/*! \brief Play a specific song (DUMB)
+/*! \brief Play a specific song (SDL2_Mixer)
  *
  * This will stop any currently played song, and then play
  * the requested song.  Based on the extension given, the appropriate player
  * is called.
+ * NOTE: position is not implemented but it is currently only ever called
+ * with value 0 anyway.
  *
  * \param   music_name The relative filename of the song to be played
  * \param   position The position of the file to begin at
  */
-void KMusic::play_music(const std::string& music_name, long position)
+void KMusic::play_music(const std::string& music_name, long)
 {
-    if (is_sound != 0)
+    if (is_sound == 0)
     {
-        const std::string fstr = kqres(MUSIC_DIR, music_name);
-        const char* filename = fstr.c_str();
-
-        stop_music();
-        if (Disk.exists(filename))
-        {
-            if (strstr(filename, ".mod"))
-            {
-                mod_song[current_music_player] = dumb_load_mod(filename);
-            }
-
-            else if (strstr(filename, ".xm"))
-            {
-                mod_song[current_music_player] = dumb_load_xm(filename);
-            }
-
-            else if (strstr(filename, ".s3m"))
-            {
-                mod_song[current_music_player] = dumb_load_s3m(filename);
-            }
-
-            else
-            {
-                mod_song[current_music_player] = NULL;
-            }
-            if (mod_song[current_music_player])
-            {
-                /* ML: we should (?) adjust the buffer size after everything is running
-                 * smooth */
-                float vol = float(gmvol) / 250.0f;
-                /* TODO
-                        mod_player[current_music_player] =
-                            al_start_duh(mod_song[current_music_player], 2, position, vol, 4096 * 4, 44100);
-                */
-            }
-            else
-            {
-                TRACE(_("Could not load %s!\n"), filename);
-            }
-        }
-        else
-        {
-            mod_song[current_music_player] = NULL;
-        }
+        return;
     }
+    if (music)
+    {
+        Mix_FadeOutMusic(1000);
+        while (!Mix_PlayingMusic())
+        {
+            Game.ProcessEvents();
+        }
+        music = nullptr;
+    }
+    Mix_PlayMusic(music_cache.get(music_name), -1);
 }
 
-/*! \brief Stop the music (DUMB)
+/*! \brief Stop the music  (SDL2_Mixer))
  *
  * Stops any music being played. To start playing more music, you
  * must call play_music(), as the current music player will no longer
@@ -179,17 +144,15 @@ void KMusic::play_music(const std::string& music_name, long position)
  */
 void KMusic::stop_music(void)
 {
-    if (is_sound != 0 && mod_player[current_music_player])
-    { /* TODO
-         al_stop_duh(mod_player[current_music_player]);
-         unload_duh(mod_song[current_music_player]);
-      */
-        mod_player[current_music_player] = NULL;
-        mod_song[current_music_player] = NULL;
+    if (is_sound == 0)
+    {
+        return;
     }
+    Mix_HaltMusic();
+    music = nullptr;
 }
 
-/*! \brief Pauses the current music file (DUMB)
+/*! \brief Pauses the current music file  (SDL2_Mixer))
  *
  * Pauses the currently playing music file. It may be resumed
  * by calling resume_music(). Pausing the music file may be used
@@ -197,48 +160,68 @@ void KMusic::stop_music(void)
  */
 void KMusic::pause_music(void)
 {
-    if (is_sound != 0)
+    if (is_sound == 0)
     {
-        if (current_music_player < MAX_MUSIC_PLAYERS - 1)
-        {
-            /* TODO
-                  al_pause_duh(mod_player[current_music_player]);
-            */
-            current_music_player++;
-        }
-        else
-        {
-            TRACE(_("reached maximum levels of music pauses!\n"));
-        }
+        return;
     }
+    Mix_PauseMusic();
 }
 
-/*! \brief Resume paused music (DUMB)
+/*! \brief Resume paused music  (SDL2_Mixer))
  *
  * Resumes the most recently paused music file. If a call to
  * play_music() was made in between, that file will be stopped.
  */
 void KMusic::resume_music(void)
 {
-    if (is_sound != 0 && current_music_player > 0)
+    if (is_sound == 0)
     {
-        stop_music();
-        current_music_player--;
-        /* TODO
-            al_resume_duh(mod_player[current_music_player]);
-        */
+        return;
     }
+    Mix_ResumeMusic();
 }
 
 void KMusic::play_effect(int, int)
 {
 }
-void KMusic::play_sample(void*, int, int, int, int)
+void KMusic::play_sample(void* chunk, int, int, int, int)
 {
+    if (is_sound == 0)
+    {
+        return;
+    }
+    Mix_PlayChannel(-1, reinterpret_cast<Mix_Chunk*>(chunk), 0);
 }
-void KMusic::set_volume(float, int)
+/* \brief Set overall sound volume.
+ * \param   volume 0 (silent) to 250 (loudest)
+ */
+void KMusic::set_volume(int sound_volume)
 {
+    dvol = float(std::clamp(sound_volume, 0, 250)) / 250.0f;
+    Mix_Volume(-1, int(128.0f * dvol));
+    Mix_VolumeMusic(int(dvol * mvol * float(MIX_MAX_VOLUME)));
 }
-int _color_depth = 8;
-void* screen;
 KMusic Music;
+
+Mix_Music* Mix_MusicLoader::operator()(const std::string& music_name)
+{
+    auto mpath = kqres(eDirectories::MUSIC_DIR, music_name);
+    return Mix_LoadMUS(mpath.c_str());
+}
+Mix_Chunk* Mix_ChunkLoader::operator()(const std::string& name)
+{
+    auto path = kqres(eDirectories::DATA_DIR, name);
+    return Mix_LoadWAV(path.c_str());
+}
+void Mix_ChunkDeleter::operator()(Mix_Chunk* chunk)
+{
+    Mix_FreeChunk(chunk);
+}
+void* KMusic::get_sample(const string& s)
+{
+    return sample_cache.get(s);
+}
+void KMusic::free_samples()
+{
+    sample_cache.clear();
+}
